QDoc: Remove Section::m_obsoleteMemberMap
QDoc uses a pair of classes `Sections`/`Section` that destruct and
categorize the members of an `Aggregate`, an internal representation for
documentable elements that are composed of multiple documentable
elements, to generate a series of ordered lists in the output
documentation.

For example, the list of members of a C++ class in the reference page of
that same class.

Internally, a `Section` categorizes the "children" of an `Aggregate` by
using a series of maps, to provide an ordering, that are later
"flattened" into ordered, sequential, monodimensional-structures,
generally vectors, that are presented to the consumer of a `Section`,
mostly the `Generator`s that actually produce the output documentation,
to iterate over the categorized elements.

The keys of the internal maps are never used by consumers of a
`Section`, such that their only purpose is to internally provide an
ordering before they are "flattened".

Hence, the maps can be removed, with the intermediate flattening step,
in favor of constructing the same monodimensional-structures,
which can be sorted directly, before being presented to a consumer, to
produce the same result.

`Section::m_obsoleteMemberMap`, is one such map that is used to hold all the
elements of a destructured `Aggregate` that are deprecated, and is
now removed as it is unnecessary.

Further to removing the member variable, all insertion into the map were
converted to insertion into `Section::m_obsoleteMembers`, the flattened version
of the map.

`Section`s are statically allocated in `Sections`, which act as a
container of `Section`, and are reused between instances of `Sections`.
To support this structure, `Section` provides a `clear` method, which
cleans the internal state, such as the internal maps, to prepare the
`Section` instance for reuse.
The clearance of `m_obsoleteMemberMap` was removed from
`Section::clear`, as the member variable is not part of the internal
state anymore.

Each internal map is "flattened" in `Section::reduce`, which is called
after the `Aggregate` that a `Section` is destructuring is fully
categorized.
The code flattening `m_obsoleteMemberMap` was removed in favor of sorting
`m_obsoleteMembers` at this step, after it is fully populated.

`Section` uses a free function in "sections.cpp", `sortName`, which
produces a series of orderable strings based on a `Node`'s, the base
class for a documentable element, name, that are generally used as keys
for the map, providing the specific ordering that QDoc currently uses.

To provide the same ordering, the sorting uses a previously added
comparator that uses `sortName` as its base.

It should be noted that, while the ordering uses the same methodology,
it will not be equivalent to the one provided by the map-based
implementation.

In `Section`, not all elements would first pass by the internal map
before it was "flattened", instead having them directly inserted in the
"flattened" version, such that they would receive no sorting.

For example, elements of a `SharedCommentNode`, an internal
representation for elements that share a documentation block, would
behave this way.
As the sorting now happens directly on the "flattened"
`Section::m_obsoleteMembers`, the groups of elements that are produced
from those `SharedCommentNode` will now provide an order.

This only affects, at the current point in time, two output files,
"qkeycombination-obsolete.html" and "qlist-obsolete.html".

Change-Id: Ida4e54c594c6376f2d0ccf8781e8fc07d4a5f49f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>