QDoc: Remove ClassMapList
QDoc uses a pair of classes `Sections`/`Section` that destruct and
categorize the members of an `Aggregate`, an internal representation for
documentable elements that are composed of multiple documentable
elements, to generate a series of ordered lists in the output
documentation.

For example, the list of members of a C++ class in the reference page of
that same class.

Internally, a `Section` categorizes the "children" of an `Aggregate` by
using a series of maps, to provide an ordering, that are later
"flattened" into ordered, sequential, monodimensional-structures,
generally vectors, that are presented to the consumer of a `Section`,
mostly the `Generator`s that actually produce the output documentation,
to iterate over the categorized elements.

The keys of the internal maps are never used by consumers of a
`Section`, such that their only purpose is to internally provide an
ordering before they are "flattened".

Hence, the maps can be removed, with the intermediate flattening step,
in favor of directly constructing the same monodimensional-structures,
which can be sorted directly, before being presented to a consumer, to
produce the same result.

QML elements are structured with relatively more complexity than other
similar maps in `Section`.
Instead of directly using a map, QML `Node`s are destructures into a
lists of pairs, where each pair associates a certain QML `Node` with its
elements, the map itself.

Those are later flattened into a list of similar pairs, where the second
element is a vector of the "values" of the map instead of the map itself.

Supposedly, this is used to support the generation of "all members"
pages for QML elements.
Those pages differ from the one generated for C++ classes as they group
the various members based on the QML element that "generates" them, that
is, inherithed elements are group under the element they are inherited
from, instead of the "elements soup" that the C++ "all members" pages
use.
Hence, the pairs structure simplifies this process, as it preserves a
reference to the generating member and provides the required grouping.

`Section::m_classMapList`, is the list of pairs that is used to hold
all the elements of a destructured `Aggregate` that are QML elements,
and is now removed as it is unnecessary.

Further to removing the member variable, all insertion into the list
were converted to insertion into `Section::m_classNodesList`, the
flattened version of the structure.

Those insertions were performed in two steps and through two methods,
`Section::newClassMap` and `Section::add`.

`Section::newClassMap` was used to create a new pair to add to
`m_classMapList` and is now modified to directly add a pair to
`m_classNodesList`.
`Section::newClassMap` had only a single use, in
`Section::buildStdQmlTypeRefPageSections`, where QML `Aggregate` are
destructured, and was thus removed in favor of moving the construction
of the required pairs inline.

`Section::newClassMap` would create the pairs on the heap and then store
a pointer to them in `m_classMapList`.
There is no currently known reason for this requirement, such that along
with the removal and inlining of `Section::newClassMap` the pairs were
modified to be stored directly.

As the pairs are not on the heap anymore, `ClassNodesList`, the type of
`m_classNodesList` was modified to be a list of `ClassNodes`, a typedef
for the stored pairs, instead of a list of pointers to `ClassNodes`.
The single consumer of `m_classNodesList`,
`HtmlGenerator::generateAllQmlMembersFile`, which generates the actual
"all members" file for QML elements, was modified to take into account
the change in type.

`Section::add` would store a new element in the map of the last added
pair, to populate it, while ensuring that the element was added to the
"all members" category too.
Since it had a single usage, it was removed in favor of an equivalent
inline version.

`Section`s are statically allocated in `Sections`, which act as a
container of `Section`, and are reused between instances of `Sections`.
To support this structure, `Section` provides a `clear` method, which
cleans the internal state, such as the internal maps, to prepare the
`Section` instance for reuse.
The clearance of `m_classMapList` was removed from `Section::clear`, as the
member variable is not part of the internal state anymore.

Furthermore, due to the removal of the heap allocations for the stored
pairs, calls to `qDeleteAll` to free the allocated memory in
`Section::clear` were removed.

As `Section` destructures an `Aggregate` into categories, not all of
which might be used depending on the structure of the specific
`Aggregate`, it provides a custom `isEmpty` method that specifies that
the `Section` is empty only if all of the internal containers for each
category is empty.
This method was changed to use `m_classNodesList` instead of
`m_classMapList`, as `m_classMapList` is now removed in favor of
`m_classNodesList`.

Each internal map is "flattened" in `Section::reduce`, which is called
after the `Aggregate` that a `Section` is destructuring is fully
categorized.
The code flattening all maps in `m_classMapList` was removed in favor of
sorting all vectors of `m_classNodesList` at this step, after it is
fully populated.

`Section` uses a free function in "sections.cpp", `sortName`, which
produces a series of orderable strings based on a `Node`'s, the base
class for a documentable element, name, that are generally used as keys
for the map, providing the specific ordering that QDoc currently uses.

To provide the same ordering, a previously added comparator that uses
`sortName` as its base was used.

It should be noted that, while the ordering uses the same methodology,
it will not be equivalent to the one provided by the map-based
implementation, s `sortName` produces colliding keys for certain
different elements.
The ordering for such colliding elements is implementation dependent and
is inverted between the map-based version, which orders colliding
elements in reversed insertion order, and the "flattened" version, which
orders colliding elements in insertion order.

This affects many files in the output documentation, generally the
"members" file that QDoc produces for a QML type, that lists all
elements of the type, inherited or not.

Change-Id: I6ad888af00d83cb3bc6c7a1a852718a7c5e207ad
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>