qdoc: Clean up methods related to tags in ManifestWriter
Tags are specific to each example, so having the "current" list
of tags as a class member for ManifestWriter is misleading. Instead,
adapt the signatures of the helper methods so that tags are explicitly
passed as arguments and return values.

Now that the methods do not access object state anymore, they can also
be static functions.

Pick-to: 6.4
Change-Id: I7edceed18abf24c3c79f6b44d4ac11101ec307ee
Reviewed-by: Luca Di Sera <luca.disera@qt.io>