QDoc: Remove Section::Status
QDoc uses a pair of classes, `Sections`/`Section`, to categorize certain
`Aggregate`s, the internal repreentation for documentable elements
composed of multiple documentable elements, to generate a series of
ordered and organized lists in the generated output documentation for
that same `Aggregate`.

For example, it uses those classes to categorize the members of a C++
class so as to build a list of all members of that same class.

A `Section` has a series of internal storages for certain categories of
elements.
For example, it separately stores elements that are
"obsolete"/"deprecated".

Furthermore, it has a series of parameters that are used during the
generation of the output documentation to branch the required style and
processing for the generated elements.

One such parameter is the "status" of the `Section`, set at construction
time, which uses an enumeration, `Section::Status`, with two possible
values, `Obsolete` and `Active`.

Currently, all usages of `Section` always set the "status" to `Active`,
such that the field is redundant.
Internally, no sensible usage of the field or of the enumeration itself
are performed.

The original meaning of the field is unknown. A `Section`, treats
"obsolete" element separately, albeit it always provides storage and
access to the categorization of all elements at the same time,
"obsolete" or not.
Supposedly, it might have originally been
intended to specifically mark `Section` instances that were intended to
deal with "obsolete" members only.

Nonetheless, as the field is unrequired, it is now removed as dead code.

`Section::m_status`, the member storing the status of a `Section` was
removed.
Its getter, `Sections::status`, was removed as a consequence.

Additionally, the only constructor of `Section` was modified to not
require a `Section::Status` argument and to not set `m_status`, as a
consequence of the removal.

Each call of the constructor was updated to remove the usage of the
`Status` argument.

The `Section::Status` enumeration was hence removed, as it generally
lost its purpose.

`Section::insertReimplementedMember`, a method that took care of
dispatching an element that was categorized as reimplemented, e.g an
overridden virtual function, was modified to remove its single usage of
`Section::m_status` and `Section::status`.

The method branched on the condition of `m_status` being `Active`. As
all `Section` used in QDoc have an `Active` status, the condition was
unnecessary and tautological in the current codebase.

A part of the documentation of `Section::insertReimplementedMember` was
removed as it referred to the now removed condition.

A method, `generateSectionList`, present on both the `HtmlGenerator` and
`DocbookGenerator` classes, used to generate a list in the output
documentation based on the contents of a `Section`, took a `Status`
parameter whose only purpose was to decide which storage in `Section` to
use.

That is, the method would consider the "contents" of the `Section` to be
the categorized "obsolete" members if an `Obsolete` status was passed
and all other members otherwise.

The parameter was defaulted so that `Active` was the default choice and
only a single call of the method specified the parameter, passing an
`Obsolete` status.

To allow the removal of `Section::Status`, a temporary change was made to the
interface of the method, so that it uses a `bool` parameter in place of
the original `Status` one.
When the parameter is `false`, its default value, the method behaves as
if an `Active` status was passed, behaving as if `Obsolete` was passed.

Both declaration, definition and usages of the method were updated to
conform to the new interface.

While the usage of a boolean is both more restrictive, extension-wise,
and less clean, all the code related to the `Generator` classes has to
be revised and, almost certainly, changed substantially, such that this
"uglification" of the code is to be considered only temporary, allowing
it to be ignored for the time being while still enabling a
simplification in the too-complex code for `Section`.

Albeit unrelated, the documentation for a now non-existing constructor
of `Section` was removed from "sections.cpp".

Change-Id: Ie234f58f36e99ced40574440d51a9f89ddefb1da
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>