qdoc: Avoid invalid links generated for QML value types
In .index files, a 'qml-module-name' attribute was incorrectly added
for QML (value) types that did not use the \inqmlmodule command.
Instead of omitting the attribute, its value was set to the name of
the type itself. This caused incorrect linking to
"<type>-qmlmodule.html" in some cases.

Drop the attribute for QML types that do not provide the module
information. Also drop it for QML modules as it's duplicate
information to the 'name' attribute.

Add a test for ensuring that autolinking to QML types loaded
from .index works correctly.

Change-Id: Idb00999774eb2411f406148011e4542d8038c6cd
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
(cherry picked from commit 404821bbdee8d885d60b18d3b11ad11b2bcdfd30)
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>