CMake: Fix DesignerComponentsPrivate private headers to be synced
Add lib_pch.h to the list of headers so it is synced by syncqt cpp as
a private header. This doesn't seem right, but it matches what we did
in < Qt 6.4 as well as Qt 5.15.

Without at least one synced private header, the build system code
generates incorrect INTERFACE include directories entries which fail
configuration of user projects when the target in question is linked.

Pick-to: 6.5
Fixes: QTBUG-109423
Change-Id: I11b48954044f057cb66bfc74c34dde5e1ba5739d
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>