QDoc: Fix documentation and configuration warnings about libclang
The most important thing a user needs to know when installing libclang
on linux is that they need the ClangConfig.cmake file. State that when
we fail to find libclang.

Also, add a note about LLVM_INSTALL_DIR to the documentation. Using
LLVM_INSTALL_DIR is the cleaner option since the configure script should
only look for LLVM-related packages in the LLVM installation.

Change-Id: Iaa26a4e15955b60120f002a34ab8d7a00548a221
Reviewed-by: Paul Wicking <paul.wicking@qt.io>