QDoc: Tag C++ elements as "explicit" when they are marked as such
C++'s constructors and conversion functions can be marked by
the explicit specifier, so that the element cannot participate in
implicit conversion and copy-initialization.

This information is important to the user of an API, as it may dictate
when and how they can use the given constructor or conversion function.
Nonetheless, up to now, QDoc would not treat the "explicit" specifier,
such that this information was lost between source-code and
documentation.

To avoid the issue and ensure that users are given a way to discern
whether something is marked explicit or not, QDoc will now "tag" a
relevant documentable element as "explicit" when the "explicit"
specifiers appears in its declaration.

To store this information, `FunctionNode`, the internal representation
for "callable" documentable elements, such as a C++ constructor or
method, was modified to expose a boolean flag, `m_explicit`, trough the
a setter, `markExplicit`, which enables the flag, and a getter,
`isExplicit`, which allows to peek into the flag value.

The flag is initially defaulted to `false` and is set only when the
"explicit" specifier is encountered in the declaration of the relevant
element.

Generally, this kind of information is extracted from the codebase in
`ClangCodeParser::processFunction`, where an instance of a
`FunctionNode` is populated based on the Libclang-provided `CXCursor`
that generated it.

As Libclang does not have direct support for the `explicit` specifier,
at the current point in time, we extract the information based on the
tokenized version of the relevant declaration.

That is, the source code represented by the relevant `CXCursor`, which
will generally represent a declaration, is
tokenized by Clang, all keywords tokens that are part of the declaration
are inspected and, if an "explicit" keyword is encountered, we consider
the given declaration to be marked "explicit".

To do so, a static function `get_specifiers` was introduced in
"clangcodeparser.cpp", which performs the above process.

The function returns a newly introduced structure, `CXXSpecifiers`, that
contains some minimum information with regards to the specifiers that we
are interested into that Libclang does not directly support, such as
"explicit".

`ClangCodeParser::processFunction` will now call this newly introduced
function and mark any relevant information in the processed
`FunctionNode` based on the returned `CXXSpecifiers` instance.

Later on, during the "Generation Phase", where QDoc destructures the
provided documentable elements to produce the final output
documentation, QDoc will call a `CodeMarker` to generate the set of
"tags", small strings that appear near the detailed documentation of an
element, to enhance the final documentation with certain information
about the documented element.

To make use of the now stored "explicit" information,
`CodeMarker::extraSynopsis`, which is the particular methods that
generated the relevant set of "tags" for an element, was modified to
take into account the "explicit" flag of a `FunctionNode`, and generate
an "explicit" "tag" if it is set to `true`.

When QDoc parses the source code for a project, it generates an
XML-based "index" file, containing certain information about the
extracted documentable elements that can be consumed by external tools
and is consumed by QDoc itself to enable cross-module linking.

To allow the newly added "explicit" information to be retained between
modules and to enable the inspection of such information for external
consumers, `QDocIndexFiles::generateFunctionSection`, which writes the
relevant information about a `FunctionNode` in the index file, was
modified to write an "explicit" attribute in the XML-element
representing the `FunctionNode`.

The attribute is elided when the `FunctionNode` is nor marked as
"explicit", the default state, to save space and avoid cluttering the
output.

Similarly, `QDocIndexFiles::readIndexSection`, which retrieves the
information stored in a certain index file and rebuilds the internal
representations that QDoc uses for the represented elements, was
modified to read the "explicit" attribute that was added.

If the attribute is present and has a "true" value in a "function"
element, the reconstructed `FunctionNode` will be marked as "explicit",
to retain the information.

Fixes: QTBUG-109370
Pick-to: 6.5
Change-Id: Iab86082cc63d1191024c514bd197aade63e1d4f6
Reviewed-by: Luca Di Sera <luca.disera@qt.io>