QDoc: Tag C++ elements as "constexpr" when they are marked as such
C++'s allows certain declarations to be marked by
the `constexpr` specifier, so as to signify that the element can be used
in constant expressions contexts.

This information is important to the user of an API, as it may dictate
when and how they can use a certain element.
Nonetheless, up to now, QDoc would not treat the "constexpr" specifier,
such that this information was lost between source-code and
documentation.

To avoid the issue and ensure that users are given a way to discern
whether something is marked "constexpr" or not, QDoc will now "tag" a
relevant documentable element that is a callable, such as a C++ method,
as "constexpr" when the `constexpr` specifier appears in its
declaration.

To store this information, `FunctionNode`, the internal representation
for "callable" documentable elements, such as a C++ constructor or
method, was modified to expose a boolean flag, `m_constexpr`, trough the
a setter, `markConstexpr`, which enables the flag, and a getter,
`isConstexpr`, which allows to peek into the flag value.

The flag is initially defaulted to `false` and is set only when the
"constexpr" specifier is encountered in the declaration of the relevant
element.

Generally, this kind of information is extracted from the codebase in
`ClangCodeParser::processFunction`, where an instance of a
`FunctionNode` is populated based on the Libclang-provided `CXCursor`
that generated it.

As Libclang does not have direct support for the `constexpr` specifier,
at the current point in time, we extract the information based on the
tokenized version of the relevant declaration.

That is, the source code represented by the relevant `CXCursor`, which
will generally represent a declaration, is
tokenized by Clang, all keywords tokens that are part of the declaration
are inspected and, if a "constexpr" keyword is encountered, we consider
the given declaration to be marked "constexpr".

To do so, we reuse the infrastructure that was introduced in a recent
commit to preserve the information for `explicit` specifiers.

The static function `get_specifiers` in
"clangcodeparser.cpp", performs the above process, and is now modified
to take into account the `constexpr` specifier.

The function returns a structure, `CXXSpecifiers`, that
contains some minimum information with regards to the specifiers that we
are interested into that Libclang does not directly support.
The structure is now updated to provide information about the presence,
or lack thereof, of a `constexpr` specifier.

Similarly, the body of `get_specifiers` was modified to correctly
populate this information based on the available tokens.

`ClangCodeParser::processFunction`, which calls `get_specifiers`, was
modified to take into account the new "constexpr" information in the
returned `CXXSpecifiers` value, marking the processed `FunctionNode` as
"constexpr" if a `constexpr` specifier was encountered.

Later on, during the "Generation Phase", where QDoc destructures the
provided documentable elements to produce the final output
documentation, QDoc will call a `CodeMarker` to generate the set of
"tags", small strings that appear near the detailed documentation of an
element, to enhance the final documentation with certain information
about the documented element.

To make use of the now stored "constexpr" information,
`CodeMarker::extraSynopsis`, which is the particular methods that
generates the relevant set of "tags" for an element, was modified to
take into account the "constexpr" flag of a `FunctionNode`, and generate
an "constexpr" "tag" if it is set to `true`.

When QDoc parses the source code for a project, it generates an
XML-based "index" file, containing certain information about the
extracted documentable elements that can be consumed by external tools
and is consumed by QDoc itself to enable cross-module linking.

To allow the newly added "constexpr" information to be retained between
modules and to enable the inspection of such information for external
consumers, `QDocIndexFiles::generateFunctionSection`, which writes the
relevant information about a `FunctionNode` in the index file, was
modified to write a "constexpr" attribute in the XML-element
representing the `FunctionNode`.

The attribute is elided when the `FunctionNode` is not marked as
"constexpr", the default state, to save space and avoid cluttering the
output.

Similarly, `QDocIndexFiles::readIndexSection`, which retrieves the
information stored in a certain index file and rebuilds the internal
representations that QDoc uses for the represented elements, was
modified to read the "constexpr" attribute that was added.

If the attribute is present and has a "true" value in a "function"
element, the reconstructed `FunctionNode` will be marked as "constexpr",
to retain the information.

Task-number: QTBUG-93439
Change-Id: I63b18c34d7b9f3d72047ab2ba823d105f5a694e7
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
(cherry picked from commit feb4a767b2728f706996217fd90efc54b2c3b2f1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>