Assistant: port HelpEngineWrapper from QSharedPointer to unique_ptr
The only reason why the code used a shared instead of a unique ptr,
apart from possibly predating C++11, is that the container in which
the pointer is held, QMap, doesn't support move-only types.

Now that QMap is but a wrapper around std::map, use the Real Thing™,
and gain move-only type support for free.

Replace a QPair with a proper struct with named members.

Pick-to: 6.5 6.4
Change-Id: Ic8d5529c5b03fa570553c8f81d7f9a721a167246
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>