Designer: port from QSharedPointer to unique_ptr
The PropertyHelper objects are never shared, so the only reason to use
a shared instead of a unique pointer, apart from the code possibly
predating C++11, is that the container in which they're held, QList,
cannot deal with move-only types.

Change the container to one that can (std::vector) and use unique_ptr
instead of QSharedPointer.

As a drive-by, make the createPropertyHelper() factory function return
by unique_ptr directly.

This removes the last user of QSharedPointer in this module, making it
ready for an eventual QT_NO_SHARED_POINTER.

Change-Id: I7051b351c4abbd1099db46823b52b5b4b16fdab5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 3d050f8f2dd53827a2a7cb594d42e60583902475)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>