QDoc: Remove unused overload of \meta command
At some point, special processing of meta content was added
to make QDoc generate an entry of maintainer(s) in output.
However, this hasn't seen use in the docs going back as far as
_at least_ Qt 4.8.

Remove this overload of the \meta command as it is unused,
and adjust the relevant documentation in the QDoc manual.

Fixes: QTBUG-109724
Change-Id: I9890a831f0c950c122eced62e2e1b184bda23b0b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit db435620ab5d5420d1b59f96f413425a15c3b142)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>