QDoc: Require clangcpp as a dependency
QDoc uses a clang-based parser to extract information, from C++ source
code, that is used to provide some common functionalities and safety
guarantees when documenting a C++ project.

The parser is implemented in terms of LibClang, clang's C API, a wrapper
around clang's C++ API.

LibClang is a generally stable API that exposes some basic
functionalities of the underlying C++ AST, but does not give full
control over its traversal and does not expose all the available
functionality of the underlying technology.

Certain lack of features have been proven to be a common blocker for
relatively recent user requests, especially in regards to the evolution
of the C++ language itself, which, in turn, would require an
unnecessary and error prone amount of work to be performed on our side
to provide the required support.

To address the issue, a series of patches were pushed upstream to expose
functionalities of the C++ API to LibClang to enable us to more easily
complete the relevant user requests.

While this method is relatively feasible, the development cycle remain
hindered.
All improvements that are merged upstream will not be available to us
until a new version is out. Furthermore, upgrades to new versions have
been proven difficult in the current infrastructure, as they are tied to
other projects in Qt, such as lupdate, where this coupling might slow
down due to the need to update code that is now invalid.

The method is further unreliable due to being dependent on another
project release cycle and contribution workflow.
While the upstream project has shown to be responsive to our requests,
the lack of control does not allow us, in certain cases, the required
freedom to swiftly respond to our users concerns.

To fix this issue, QDoc will at some point require a move from the
current LibClang-based parser to a new parser written using Libtooling
and the C++ API.
Nonetheless, due to the current state of the codebase and the general
lack of manpower for the maintenance and improvement of QDoc, this is
not expected to be able to be completed in a relatively short amount of
time.

To avoid the current unreliability and avoid the requirement to move
from the current parser to a completely new one, a different approach
was designed.
Using a certain, expected to be stable, almost public, implementation
detail of LibClang, we expect to be able to drop down to the C++ API,
from LibClang, on the spot, allowing us to have mostly full access to
the underlying API features.

This is further expected to allow us to slowly move certain components
and processes to the C++ API, incrementally, while possibly satisfying
users' requests, without the upfront work required to fully move away
from the current parser.

As we expect to start using this pattern and for it to become more
common in the new future, a dependency on the "clangcpp" library was
added to QDoc, which will fail to configure should it be missing.

The new dependency, while still more restricting than before, is
generally expected to not be disruptive to most users when upgrading, as
our prepackaged binaries already provide both libraries and the users
who prefer to maintain their own system library should be able to easily
reach for the C++ library if they have LibClang available, which depends
on the C++ library in the first place.

The help message that is reported when QDoc is not able to be configured
was modified to not refer directly to LibClang and more correctly refer
to the requirement of both the C and C++ libraries to be available.

A line suggesting relatively unspecific, Linux-only, specific-distros only, ways of
installing the required libraries was removed as it serves only users
who use the related distros, with not enough precision, which are not
proven to represent any major chunk of our user-base.
A similar line that is macOS only was not removed with the same
reasoning as it has higher precision.

[ChangeLog][QDoc] QDoc will now require "clangcpp" to be available.

Task-number: QTBUG-93439
Change-Id: I38ffb5a64501bf1c4726b6fb09758fd1cc0f2c3c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>