QDoc: Use clang's C++ API to mark element constexpr
QDoc uses clang's C API, LibClang, to parse the source code of C++
projects, so as to be able to extract the user-written documentation and
provide certain guarantees about its correctness.

When an element is marked as "constexpr", QDoc extracts that information
and provides that information in the produced output documentation in a
user-visible way.

As LibClang does not provide first party support for extracting
"constexpr" information, a workaround was provided to perform this
operation where the source code of the processed element was tokenized
and the element was marked "constexpr" if a "constexpr" keyword appeared
within its tokens.

This approach is feeble. Indeed, the code that performed this process
was bugged.
For methods/functions whose definition was inline, the tokenized source
code would be contain the body, where the "constexpr" keyword can appear
in non-specifier contexts (e.g `if constexpr`), resulting in false
positives being marked as "constexpr".

Similarly, certain elements might be implicitly "constexpr" even if no
`constexpr` specifier appears in their declaration/definition, such as
certain defaulted constructors.

Recent changes to QDoc allows us to turn to clang's C++ API, which is more
complete, on the spot from a LibClang's cursor.
The C++ API has first-party support for "constexpr" detection and can
thus ensure that all relevant cases are taken care of correctly.

Hence, the "constexpr" detection method was revamped in term of the C++
API, resolving the above mentioned issues.

`ClangCodeParser::processFunction` is the method that generally takes
care of extracting this form of information, enriching a `FunctionNode`,
the internal representation for documentable elements that are callable,
with the retrieved information.

`ClangCodeParser::processFunction` would call the free-function
`get_specifiers`, which performed the tokenization method to extract
specifiers-related information, producing a `CXXSpecifiers` instance,
that contains data about the presence or lack thereof of certain
specifiers which QDoc support.
The produced instance was then used to mark the processed `FunctionNode`
based on the information that were extracted.

`ClangCodeParser::processFunction` was modified to avoid processing
"constexpr"-related information from the produced `CXXSpecifiers`,
instead retrieving it trough clang's C++ API's `isConstexpr` method.

`get_specifiers` was modified so as to not extract any "constexpr"
information anymore, due to the now missing usage for that information.
`CXXSpecifiers` was in turn modified to not store any
"constexpr"-related information, for the same reason.

Any documentation that contained non-constexpr inline-methods whose body
contained the `constexpr` keyword will have its output documentation
modified by this change, removing the produced "constexpr" tag from
those method, which were false positives.

Any documentation that contained implicitly constexpr methods will have
its output documentation modified so that those methods will now be
marked as "constexpr".

Change-Id: Ib47b58db200905f95f8358bcc363ff1187f01741
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>