QDoc: Use clang's C++ API to mark element explicit
QDoc uses clang's C API, LibClang, to parse the source code of C++
projects, so as to be able to extract the user-written documentation and
provide certain guarantees about its correctness.

When an element is marked as "explicit", QDoc extracts that information
and provides it in the produced output documentation in a user-visible
way.

As LibClang does not provide first party support for extracting
"explicit" information, a workaround was provided to perform this
operation where the source code of the processed element was tokenized
and the element was marked "explicit" if an "explicit" keyword appeared
within its tokens.

This approach is feeble. Indeed, the code that performed this process
was bugged.

Certain elements whose code was generated from or contained macros might
not have been tokenized correctly as to expose an "explicit" specifiers
or might have contained non-relevant "explicit" keyword that was
misinterpreted.
Certain elements of this issue could be solved directly in LibClang,
with a more thorough manipulation of the relevant cursors before
tokenization, but would require a huge amount of work to identify and
cover various possible cases.

Recent changes to QDoc allows us to turn to clang's C++ API, which is more
complete, on the spot from a LibClang's cursor.
The C++ API has first-party support for "explicit" detection and can
thus ensure that all relevant cases are taken care of correctly.

Hence, the "explicit" detection method was revamped in term of the C++
API, resolving the above mentioned issues.

`ClangCodeParser::processFunction` is the method that generally takes
care of extracting this form of information, enriching a `FunctionNode`,
the internal representation for documentable elements that are callable,
with the retrieved information.

`ClangCodeParser::processFunction` would call the free-function
`get_specifiers`, which performed the tokenization method to extract
specifiers-related information, producing a `CXXSpecifiers` instance,
that contains data about the presence or lack thereof of certain
specifiers which QDoc support.
The produced instance was then used to mark the processed `FunctionNode`
based on the information that were extracted.

`ClangCodeParser::processFunction` was modified not call
`get_specifiers` anymore, instead retrieving it trough clang's C++ API's
`isExplicit` method.

`get_specifiers` was thus removed as it had no usage anymore.
`CXXSpecifiers` was similarly removed due to lack of usage.

Certain elements of Qt's documentation that were incorrectly marked
"explicit" will now avoid the erroneous marking.
For example, certain friend operators of `QExplicitlySharedDataPointer`,
whose source code was incorrectly tokenized due to the presence of
certain macros.

Similarly, certain elements of Qt's documentation that were not
correctly marked as "explicit" will now provide the correct user-visible
information.
For example, certain constructors of `QChar` whose "explicit" specifier
was generated by a macro.

Change-Id: I33a86a4d63c67432606c7dd442d69d6a11762c89
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>